Correct fixed costs in NCTCOG
finalizeNetwork blanks the class costs to initialize them.  But the NCTCOG network generates class costs before finalizeNetwork is called, so we need to preserve them.  This is a bit of an ugly hack... we should clean up the order of generating class costs for both NCTCOG and TNTP after the validation is done.